+++ src/components/Chat/Message/styles.ts
@@ -32,7 +32,7 @@ export const MessageAuthorName = styled.Text`
   color: ${(props) => props.theme.colors.light_heading};
 `;
 
-export const MessageContentContainer = styled.TouchableOpacity<IMessageProps>`
+export const MessageContentContainer = styled.Pressable<IMessageProps>`
   position: relative;
   background-color: ${(props) =>
     !props.isRight ? props.theme.colors.shape : props.theme.colors.primary};