+++ src/components/Chat/RecordingAudio/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import { Feather } from "@expo/vector-icons"
-import { millisToTime } from "../../../utils/format"
+import { Feather } from "@expo/vector-icons";
+import { millisToTime } from "../../../utils/format";
 import {
   RecordingAudioContainer,
   RecordingAudioWrapper,
@@ -8,22 +8,35 @@ import {
   RecordingAudioDuration,
 } from "./styles";
 import { useTheme } from "styled-components";
+import { MotiView } from "moti";
 
 interface RecordingAudioProps {
-  audioDuration: number
+  audioDuration: number;
 }
 
 const RecordingAudio = ({ audioDuration }: RecordingAudioProps) => {
-
-  const { colors } = useTheme()
+  const { colors } = useTheme();
 
   return (
     <RecordingAudioContainer>
-      <RecordingAudioWrapper>
-        <RecordingAudioText>
-          <Feather name="mic" size={20} color={colors.red} /> Gravando
-        </RecordingAudioText>
-      </RecordingAudioWrapper>
+      <MotiView
+        from={{
+          opacity: 1,
+        }}
+        animate={{
+          opacity: 0.1,
+        }}
+        transition={{
+          type: "timing",
+          duration: 800,
+        }}
+      >
+        <RecordingAudioWrapper>
+          <RecordingAudioText>
+            <Feather name="mic" size={20} color={colors.red} /> Gravando
+          </RecordingAudioText>
+        </RecordingAudioWrapper>
+      </MotiView>
       <RecordingAudioDuration>
         {millisToTime(audioDuration)}
       </RecordingAudioDuration>